tg-bot: refactoring, send updates nil fix(a.bot was provided after DI and this cause panic nil pointer dereference)
